+++ src/auth/authMessages.js
@@ -16,7 +16,7 @@ import { encryptECIES, decryptECIES } from '../encryption'
 
 import { Logger } from '../logger'
 
-const VERSION = '1.2.0'
+const VERSION = '1.3.0'
 
 type AuthMetadata = {
   email: ?string,
@@ -148,6 +148,8 @@ export function decryptPrivateKey(privateKey: string,
  * @param {String} transitPublicKey the public key provide by the app
  * in its authentication request with which secrets will be encrypted
  * @param {String} hubUrl URL to the write path of the user's Gaia hub
+ * @param {String} blockstackAPIUrl URL to the API endpoint to use
+ * @param {String} associationToken JWT that binds the app key to the identity key
  * @return {String} signed and encoded authentication response token
  * @private
  */
@@ -159,7 +161,9 @@ export function makeAuthResponse(privateKey: string,
                                  appPrivateKey: ?string = null,
                                  expiresAt: number = nextMonth().getTime(),
                                  transitPublicKey: ?string = null,
-                                 hubUrl: ?string = null): string {
+                                 hubUrl: ?string = null,
+                                 blockstackAPIUrl: ?string = null,
+                                 associationToken: ?string = null): string {
   /* Convert the private key to a public key to an issuer */
   const publicKey = SECP256K1Client.derivePublicKey(privateKey)
   const address = publicKeyToAddress(publicKey)
@@ -180,6 +184,8 @@ export function makeAuthResponse(privateKey: string,
       email: metadata.email ? metadata.email : null,
       profile_url: metadata.profileUrl ? metadata.profileUrl : null,
       hubUrl,
+      blockstackAPIUrl,
+      associationToken,
       version: VERSION
     }
   } else {