+++ src/operations/safety.js
@@ -130,6 +130,11 @@ function addressCanReceiveName(address: string) {
       .then(validNames => validNames.filter(nameValid => nameValid).length < 25)))
 }
 
+function isAccountSpendable(address: string, tokenType: string, blockHeight: number) {
+  return config.network.getAccountStatus(address, tokenType)
+    .then(accountStatus => accountStatus.transfer_send_block_id >= blockHeight)
+}
+
 export const safety = {
   addressCanReceiveName,
   isInGracePeriod,
@@ -140,5 +145,6 @@ export const safety = {
   isNamespaceAvailable,
   revealedNamespace,
   namespaceIsReady,
-  namespaceIsRevealed
+  namespaceIsRevealed,
+  isAccountSpendable
 }