+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [18.1.0] - 2018-10-24
 
 ### Added
 - The `BlockstackWallet` class in `blockstack.js` supports generating
@@ -16,6 +16,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - In the transaction generation library, the makeTX functions now
   take an optional 'buildIncomplete' argument, allowing you to get
   a serialized transaction which hasn't been fully signed yet.
+- A `blockstackAPIUrl` field to the authentication response token that overrides the default
+  in `blockstack.config.network` allowing the user to specify their own Blockstack
+  Core node.
+- A `gaiaAssociationToken` field to the authentication response token which enables
+  users to run private Gaia hubs without authorization each application address.
+- An option `contentType` to the `putFile` `options` object that sets the
+  Content-Type header for unencrypted data. Thanks to @muneebm for this!
 
 ### Changed
 - Fixed a bug in version checking during the authentication process
@@ -31,6 +38,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   version 2 transactions, which breaks our interoperability with a handful
   of other libraries. Finally, with this comes a little bit of refactoring,
   to reduce the repeated code in the transaction libraries.
+- Increments the authentication process version to 1.3 in a backwards compatible change.
+- When using the bitcoind client in development networks,
+  track which addresses we've already called `importaddress` with
+  and do not retry.
 
 ## [18.0.4] - 2018-08-06